depth map dataloading debug

+++ mmseg/datasets/pipelines/formating.py
@@ -95,7 +95,7 @@ class ImageToTensor(object):
             if len(img.shape) < 3:
                 img = np.expand_dims(img, -1)
             img = img.transpose(2, 0, 1)
-            mask = np.expand_dims(results["vanishing_mask"], axis=0)
+            mask = np.expand_dims(results["vanishing_mask"], axis=0).astype(img.dtype)
             img = np.concatenate([img, mask], axis=0)
             results[key] = to_tensor(img)
         return results
